Create README.md
The Electricity Billing System is an admin side project built to ease the overall process of bill generation from the admin's side. Admins could include the electricity billing officers or the employees who generate the electricity bill for the customers. The main intention of the project is to make the process more efficient than it actually is currently since the process becomes fully computerized after we implement the project. 
For developing the same, Java programming language along with Structured Query Language, popularly known as SQL has been used. 
The overall process is as follows:-
1) Initially, there would be a login page using which the admin has to log in. The login details are verified through the details which would be stored in the table which would, in turn, be stored in our database which we would be creating using SQL.
2) After successful logging, the admin would be directed to the home page.
3) Home page contains various sections which are as follows:-
1*] Master section- It contains 2 options- New Customer and Customer Details. 
New Customer would be used if the customer has to create his account, so the admin would essentially be filling some necessary details in the form which would pop up after clicking the New Customer option and all the details would then be stored in the Customer table which would have been created for storing the new customer details.
Customer Details would just be used by the admin to display the details which have been filled in the Customer table.
2*] User section contains a Pay Bill option which would be used by the admin to pay the bill of the respective customer and the Bill details would be stored in another table. 
3*] Report section which contains a Generate Bill option which could help the admin to generate bills for that respective customers.
4*] Exit section as the name suggests would be used to exit the homepage along with the application.