Add files via upload
This file contains code required to make a connection with the database in which different tables are stored on which various queries such as fetching are performed. Using the object of this class, we can make a connection with our database each time we require and since we do not have to write the entire code again and just use the object of the connection class, it essentially saves time.