Add files via upload
This file contains the code required to create a new customer. So, essentially, using this code, a small pop-up arises after clicking the new customer option and the admin has to fill in the details of the customer after which the details are stored in the database, hence creating a new customer.