Add files via upload
Simple code whose sole motive is to help the admin view all the customer details. Here, we just fetch all the details using the SELECT query after which all the details which are stored in the customer table are fetched and displayed on the screen in the form of a table. Code also contains the facility to print pr download the customer details on his/her local printer or laptop respectively.