Add files via upload
Code which would help admins pay bills of their customers after which the admin can generate the bill and the process could be completed. The code essentially asks for the meter number, month, and the number of units consumed after which the details are stored in another table which is fetched while generating electricity bills.