Add files via upload
Code which would help admins generate bills using meter number and month. Similar details were entered while we were in the pay bill and add customer scenarios, hence details corresponding to those specific meter numbers would be fetched and displayed on the screen after which the admin would print the bill on his/her printer/laptop and give it to the customer.